Extract ClickPipes API models
